fix: Avoid leaking memory in chunk build buffers, better allocation tracking
JEmalloc makes it very easy (and performant) to shrink the allocated
storage for a VertexBufferBuilder, so we should simply reset it back
to the initial capacity when initializing the buffer. This avoids
a memory leak that could occur after rendering very large chunks, which
was caused by the buffers never being resized back down.